Hack fix to increase date precision
Since discord.py does not have a task scheduler based on a date, we makeshift our own. It will loop every second (which only actually starts that timer once the function exits. The loop will not run concurrently) and if the date equals the currently know date state then await the difference in seconds until tomorrow, then exit. When the date does not match it will instead change the channel topic and then set the state date to the current day, and begin again.